Rename `.jsx` to `.js` to have one consistent file ending
