+++ app/scripts/AddTrackPositionMenu.js
