+++ app/scripts/CollapsePanel.js
