+++ app/scripts/ConfigViewMenu.js
