+++ app/scripts/PlotTypeChooser.js
