+++ app/scripts/SearchableTiledPlot.js
