+++ app/scripts/SketchInlinePicker.js
