+++ app/scripts/TrackRenderer.js
